v2.0.0: Speeded up version
Several methods have been added to make the code faster. The computation of values and image rendering have been parallelized using OpenMP APIs, and only half of the values are calculated by exploiting the symmetry at y=0 of the fractal, effectively halving the computation time. This comes at the expense of the image rendering complexity, which starts from both the first and last rows simultaneously until reaching the center; if the number of rows is odd, the central row needs to be calculated separately, otherwise, it behaves normally. Other minor optimizations have been made, such as reducing the bit size for certain matrix types, significantly reducing the space occupied and allowing more values to be stored in the cache.